Integrate pay back method with getline.ts
Split my loan tile into separate components:
* collateral collection
* fundraising
* payback
* finished
components.
