+++ frontend/src/api/my-loans.ts
@@ -1,17 +1,18 @@
 import API, { LoanState, BigNumber } from './index';
 
 import { MyLoanT } from 'store/my-loans';
-import { getTokenSymbolsFromBlockchain, getAmountsWantedFromBlockchain, getAmountsGatheredFromBlockchain } from './utils';
+import { getTokenSymbolsFromBlockchain, getAmountsWantedFromBlockchain, getAmountsGatheredFromBlockchain, getPayBackAmountFromBlockchain } from './utils';
 
 export async function getMyLoans(): Promise<MyLoanT[]> {
   const api = await API.instance();
   let currentUser = await api.currentUser();
   let libraryLoans = await api.loansByOwner(currentUser);
   await Promise.all(libraryLoans.map(loan => loan.updateStateFromBlockchain()));
-
   const amountsGathered: BigNumber[] = await getAmountsGatheredFromBlockchain(libraryLoans);
   const amountsWanted: BigNumber[] = await getAmountsWantedFromBlockchain(libraryLoans);
   const loanTokenSymbols: string[] = await getTokenSymbolsFromBlockchain(libraryLoans);
+  const paybackAmounts: BigNumber[] = await getPayBackAmountFromBlockchain(libraryLoans);
+
   return libraryLoans.map(({
     shortId,
     description,
@@ -25,8 +26,10 @@ export async function getMyLoans(): Promise<MyLoanT[]> {
       amountWanted: amountsWanted[index],
       amountGathered: amountsGathered[index],
       tokenSymbol: loanTokenSymbols[index],
+      paybackAmount: paybackAmounts[index],
       isCollateralCollection: (loanState === LoanState.CollateralCollection),
       isFundraising: (loanState === LoanState.Fundraising),
-      isPayback: (loanState === LoanState.Payback)
+      isPayback: (loanState === LoanState.Payback),
+      isFinished: (loanState === LoanState.Finished)
     }));
 }
